requirements.txt
celery==4.1.0
sqlalchemy==1.2.8

-e ./deps/geonode
-e ./deps/django-mailer
-e ./deps/django-maploom
-e ./deps/django-osgeo-importer
-e ./deps/icon-commons

django-slack==4.1.0
ipdb==0.10.3
textile

#threaded comments dependencies
django-contrib-comments==1.6.2
django-threadedcomments==1.0.1
django-fluent-comments==1.0.5

#custom thumb dependencies
django-solo==1.1.2

# Testing dependencies
ipdbplugin==1.4.5

# Selenium testing
selenium

pandas==0.20.3

django-transaction-hooks==0.2

# CLI library used by some management commands
pick

# Change when https://github.com/garnertb/django-classification-banner/pull/6 is merged
git+https://github.com/cuttlefish/django-classification-banner.git@footer

# Pinned dependencies for geonode
beautifulsoup4==4.6.0
django-autocomplete-light==2.3.3
django-model-utils==3.0.0
elasticsearch==1.6.0
lxml==3.6.2
Pillow==5.1.0

# PKI dependencies
# for user-ordering of wildcard hostname:port mappings
django-ordered-model==1.4.3
git+https://github.com/planetfederal/django-ssl-pki
# This can be arcrest==10.3 && I think is same thing?
# Do I need these?
geonode-arcrest==10.2
# ArcREST-Package==3.5.9
# The below probably overrides the above, which is probably a dependency of above
# Something about this breaks everything where the above package don't
git+https://github.com/planetfederal/ArcREST.git@4945fd1b3325a5f1b3e7215833fd8fb1cebeb0b2#subdirectory=src
